# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "QuizApp API",
        "version": "1.0.0"
    },
    "basePath": "/part_one/public/api",
    "paths": {
        "/auth": {
            "post": {
                "summary": "Authentication Users!",
                "operationId": "authenticate",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "User object",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/me": {
            "get": {
                "summary": "Get Auth User Details",
                "operationId": "getAuthenticatedUser",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Token id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/register": {
            "post": {
                "summary": "Register Users!",
                "operationId": "register",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "User object",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
...
        "/dashboard": {
            "get": {
                "summary": "Get User Dashboard",
                "operationId": "dashboard",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Token id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/quiz/start/{id}": {
            "get": {
                "summary": "Start new Quiz",
                "operationId": "startQuiz",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Quiz id",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Token id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/quiz/myquizes": {
            "get": {
                "summary": "Get User Quizes",
                "operationId": "myQuizes",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Token id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/settings/changepassword": {
            "post": {
                "summary": "Change User Password",
                "operationId": "changepassword",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Token id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "User object",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/quiz/full/{id}": {
            "get": {
                "summary": "Full Quiz Detail",
                "operationId": "fullQuiz",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Quiz id",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Token id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "User": {
            "properties": {
                "email": {
                    "title": "Email",
                    "description": "Email",
                    "type": "string"
                },
                "password": {
                    "title": "Password",
                    "description": "Password",
                    "type": "string"
                }
            },
            "type": "object",
            "xml": {
                "name": "User"
            }
        }
    }
}